+++ src/components/Buttons/Buttons.tsx
@@ -2,6 +2,31 @@ import { type QRL, component$ } from "@builder.io/qwik";
 import { twMerge } from "tailwind-merge";
 
 export interface ButtonProps {
+  text?: string;
+  border?: string;
+  class?: string;
+  background?: string;
+  width?: string;
+  onClick$?: QRL<() => Promise<void>>;
+}
+
+export const Button = component$<ButtonProps>((props) => {
+  return (
+    <button
+      onClick$={props.onClick$}
+      class={twMerge(
+        "h-12 cursor-pointer rounded-[48px] border-2",
+        props.background,
+        props.border,
+        props.width,
+      )}
+    >
+      <div class={props.class}>{props.text}</div>
+    </button>
+  );
+});
+
+export interface ConnectButtonProps {
   image?: string;
   text?: string;
   buttonWidth?: string;
@@ -13,7 +38,7 @@ export interface ButtonProps {
   class?: string;
 }
 
-export const Button = component$<ButtonProps>((props) => {
+export const ConnectButton = component$<ConnectButtonProps>((props) => {
   return (
     <button
       onClick$={props.onClick$}
@@ -35,3 +60,5 @@ export const Button = component$<ButtonProps>((props) => {
     </button>
   );
 });
+
+
